cleaning code / updating license / add comment
