+++ packages/rum-react/src/domain/reactRouter/createRouter.spec.ts
